+++ src/controllers/validator/did-also-known-as.ts
