+++ src/controllers/validator/resource-also-known-as.ts
